requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
absl-py=2.0.0=pypi_0
accelerate=0.25.0=pypi_0
addict=2.4.0=pypi_0
aliyun-python-sdk-core=2.15.0=pypi_0
aliyun-python-sdk-kms=2.16.2=pypi_0
annotated-types=0.6.0=pypi_0
anykeystore=0.2=pypi_0
apex=0.1=pypi_0
bitsandbytes=0.42.0=pypi_0
blas=1.0=mkl
brotli-python=1.0.9=py39h6a678d5_7
bypy=1.8.5=pypi_0
bzip2=1.0.8=h7b6447c_0
c-ares=1.19.1=h5eee18b_0
ca-certificates=2023.12.12=h06a4308_0
cachetools=5.3.2=pypi_0
certifi=2023.11.17=py39h06a4308_0
cffi=1.16.0=py39h5eee18b_0
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.1.7=pypi_0
colorama=0.4.6=pyhd8ed1ab_0
conda-pack=0.7.1=pypi_0
contourpy=1.2.0=pypi_0
crcmod=1.7=pypi_0
cryptacular=1.6.2=pypi_0
cryptography=41.0.7=py39hdda0065_0
cuda-cudart=11.8.89=0
cuda-cupti=11.8.87=0
cuda-libraries=11.8.0=0
cuda-nvrtc=11.8.89=0
cuda-nvtx=11.8.86=0
cuda-runtime=11.8.0=0
curl=7.88.1=hdbd6064_2
cycler=0.12.1=pypi_0
deepspeed=0.12.6=pypi_0
defusedxml=0.7.1=pypi_0
diffusers=0.25.0=pyhd8ed1ab_0
dill=0.3.8=pypi_0
expat=2.5.0=h6a678d5_0
ffmpeg=4.3=hf484d3e_0
filelock=3.13.1=py39h06a4308_0
fonttools=4.47.0=pypi_0
freetype=2.12.1=h4a9f257_0
fsspec=2023.12.2=pyhca7485f_0
ftfy=6.1.3=pypi_0
fuser=1.0.4=pypi_0
gdbm=1.18=hd4cb3f1_4
gettext=0.21.0=h39681ba_1
giflib=5.2.1=h5eee18b_3
git=2.40.1=pl5340h36fbf9e_1
gmp=6.2.1=h295c915_3
gmpy2=2.1.2=py39heeb90bb_0
gnutls=3.6.15=he1e5248_0
google-auth=2.25.2=pypi_0
google-auth-oauthlib=1.2.0=pypi_0
greenlet=3.0.3=pypi_0
grpcio=1.60.0=pypi_0
hjson=3.1.0=pypi_0
huggingface_hub=0.20.0=pyhd8ed1ab_0
hupper=1.12.1=pypi_0
icu=73.1=h6a678d5_0
idna=3.4=py39h06a4308_0
imageio=2.34.0=pypi_0
importlib-metadata=7.0.1=pyha770c72_0
importlib-resources=6.1.1=pypi_0
intel-openmp=2023.1.0=hdb19cb5_46306
jinja2=3.1.2=py39h06a4308_0
jmespath=0.10.0=pypi_0
joblib=1.3.2=pypi_0
jpeg=9e=h5eee18b_1
kiwisolver=1.4.5=pypi_0
krb5=1.20.1=h143b758_1
lame=3.100=h7b6447c_0
lazy-loader=0.3=pypi_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libcublas=11.11.3.6=0
libcufft=10.9.0.58=0
libcufile=1.8.1.2=0
libcurand=10.3.4.101=0
libcurl=7.88.1=h251f7ec_2
libcusolver=11.4.1.48=0
libcusparse=11.7.5.86=0
libdeflate=1.17=h5eee18b_1
libedit=3.1.20221030=h5eee18b_0
libev=4.33=h7f8727e_1
libevent=2.1.12=hdbd6064_1
libffi=3.4.4=h6a678d5_0
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libiconv=1.16=h7f8727e_2
libidn2=2.3.4=h5eee18b_0
libnghttp2=1.57.0=h2d74bed_0
libnpp=11.8.0.86=0
libnvjpeg=11.9.0.86=0
libpng=1.6.39=h5eee18b_0
libssh2=1.10.0=hdbd6064_2
libstdcxx-ng=11.2.0=h1234567_1
libtasn1=4.19.0=h5eee18b_0
libtiff=4.5.1=h6a678d5_0
libunistring=0.9.10=h27cfd23_0
libwebp=1.3.2=h11a3e52_0
libwebp-base=1.3.2=h5eee18b_0
libxml2=2.10.4=hf1b16e4_1
lpips=0.1.4=pypi_0
lz4-c=1.9.4=h6a678d5_0
markdown=3.5.1=pypi_0
markdown-it-py=3.0.0=pypi_0
markupsafe=2.1.1=py39h7f8727e_0
matplotlib=3.8.2=pypi_0
mdurl=0.1.2=pypi_0
mkl=2023.1.0=h213fc3f_46344
mkl-service=2.4.0=py39h5eee18b_1
mkl_fft=1.3.8=py39h5eee18b_0
mkl_random=1.2.4=py39hdb19cb5_0
mmcv=2.1.0=pypi_0
mmcv-full=1.7.2=pypi_0
mmengine=0.10.3=pypi_0
model-index=0.1.11=pypi_0
mpc=1.1.0=h10f8cd9_1
mpfr=4.0.2=hb69a4c5_1
mpmath=1.3.0=py39h06a4308_0
multiprocess=0.70.16=pypi_0
ncurses=6.4=h6a678d5_0
nettle=3.7.3=hbbd107a_1
networkx=3.1=py39h06a4308_0
ninja=1.11.1.1=pypi_0
numpy=1.26.2=py39h5f9d8c6_0
numpy-base=1.26.2=py39hb5e798b_0
nvidia-cublas-cu12=12.1.3.1=pypi_0
nvidia-cuda-cupti-cu12=12.1.105=pypi_0
nvidia-cuda-nvrtc-cu12=12.1.105=pypi_0
nvidia-cuda-runtime-cu12=12.1.105=pypi_0
nvidia-cudnn-cu12=8.9.2.26=pypi_0
nvidia-cufft-cu12=11.0.2.54=pypi_0
nvidia-curand-cu12=10.3.2.106=pypi_0
nvidia-cusolver-cu12=11.4.5.107=pypi_0
nvidia-cusparse-cu12=12.1.0.106=pypi_0
nvidia-nccl-cu12=2.18.1=pypi_0
nvidia-nvjitlink-cu12=12.3.101=pypi_0
nvidia-nvtx-cu12=12.1.105=pypi_0
oauthlib=3.2.2=pypi_0
opencv-python=4.9.0.80=pypi_0
opendatalab=0.0.10=pypi_0
openh264=2.1.1=h4ff587b_0
openjpeg=2.4.0=h3ad879b_0
openmim=0.3.9=pypi_0
openssl=3.0.12=h7f8727e_0
openxlab=0.0.37=pypi_0
ordered-set=4.1.0=pypi_0
oss2=2.17.0=pypi_0
packaging=23.2=pyhd8ed1ab_0
pandas=2.2.1=pypi_0
pastedeploy=3.1.0=pypi_0
pbkdf2=1.3=pypi_0
pcre2=10.42=hebb0a14_0
peft=0.7.0=pypi_0
perl=5.34.0=h5eee18b_2
pillow=10.0.1=py39ha6cbd5a_0
pip=23.3.1=py39h06a4308_0
plaster=1.1.2=pypi_0
plaster-pastedeploy=1.0.1=pypi_0
platformdirs=4.2.0=pypi_0
prettytable=3.10.0=pypi_0
protobuf=4.23.4=pypi_0
psutil=5.9.7=pypi_0
py-cpuinfo=9.0.0=pypi_0
pyasn1=0.5.1=pypi_0
pyasn1-modules=0.3.0=pypi_0
pycparser=2.21=pyhd3eb1b0_0
pycryptodome=3.20.0=pypi_0
pydantic=2.5.3=pypi_0
pydantic-core=2.14.6=pypi_0
pygments=2.17.2=pypi_0
pynvml=11.5.0=pypi_0
pyopenssl=23.2.0=py39h06a4308_0
pyparsing=3.1.1=pypi_0
pyramid=2.0.2=pypi_0
pyramid-mailer=0.15.1=pypi_0
pysocks=1.7.1=py39h06a4308_0
python=3.9.18=h955ad1f_0
python-dateutil=2.8.2=pypi_0
python3-openid=3.2.0=pypi_0
python_abi=3.9=2_cp39
pytorch-cuda=11.8=h7e8668a_5
pytorch-mutex=1.0=cuda
pytz=2023.4=pypi_0
pyyaml=6.0=py39hb9d737c_4
readline=8.2=h5eee18b_0
regex=2023.10.3=py39h5eee18b_0
repoze-sendmail=4.4.1=pypi_0
requests=2.28.2=pypi_0
requests-oauthlib=1.3.1=pypi_0
requests-toolbelt=1.0.0=pypi_0
rich=13.4.2=pypi_0
rsa=4.9=pypi_0
safetensors=0.4.0=py39ha89cbab_0
scikit-image=0.22.0=pypi_0
scikit-learn=1.4.1.post1=pypi_0
scipy=1.11.4=pypi_0
setuptools=60.2.0=pypi_0
six=1.16.0=pypi_0
sqlalchemy=2.0.29=pypi_0
sqlite=3.41.2=h5eee18b_0
sympy=1.12=py39h06a4308_0
tabulate=0.9.0=pypi_0
tbb=2021.8.0=hdb19cb5_0
tensorboard=2.15.1=pypi_0
tensorboard-data-server=0.7.2=pypi_0
termcolor=2.4.0=pypi_0
thop=0.1.1-2209072238=pypi_0
threadpoolctl=3.4.0=pypi_0
tifffile=2024.2.12=pypi_0
timm=0.9.12=pypi_0
tk=8.6.12=h1ccaba5_0
tmux=3.3a=h5eee18b_1
tokenizers=0.15.0=pypi_0
tomli=2.0.1=pypi_0
torch=2.1.2=pypi_0
torch-fidelity=0.3.0=pypi_0
torchaudio=2.0.0=py39_cu118
torchvision=0.16.2=pypi_0
tqdm=4.65.2=pypi_0
transaction=4.0=pypi_0
transformers=4.36.2=pypi_0
translationstring=1.4=pypi_0
triton=2.1.0=pypi_0
typing-extensions=4.7.1=py39h06a4308_0
typing_extensions=4.7.1=py39h06a4308_0
tzdata=2024.1=pypi_0
urllib3=1.26.18=py39h06a4308_0
velruse=1.1.1=pypi_0
venusian=3.1.0=pypi_0
wcwidth=0.2.12=pypi_0
webob=1.8.7=pypi_0
werkzeug=3.0.1=pypi_0
wheel=0.41.2=py39h06a4308_0
wtforms=3.1.2=pypi_0
wtforms-recaptcha=0.3.2=pypi_0
xz=5.4.5=h5eee18b_0
yaml=0.2.5=h7f98852_2
yapf=0.40.2=pypi_0
zipp=3.17.0=pyhd8ed1ab_0
zlib=1.2.13=h5eee18b_0
zope-deprecation=5.0=pypi_0
zope-interface=6.2=pypi_0
zope-sqlalchemy=3.1=pypi_0
zstd=1.5.5=hc292b87_0
